moved stuff to their own folders and fixed imports
