+++ .gitignore
@@ -1,3 +1,4 @@
 .mypy_cache/*
 server.log
 __pycache__/*
+*.pyc
