+++ handlers/http_handlers.py
@@ -1,10 +1,13 @@
-
-from utils import HttpRequest, HttpResponse, Range
 import pathlib
 from typing import Any, List, Dict, Union, Sequence, Tuple, Callable
 import socket
 import time
 import random
+import os,sys
+BASE_DIR = os.path.join( os.path.dirname( __file__ ), '..' )
+sys.path.append(BASE_DIR)
+from utils.general_utils import HttpRequest, HttpResponse, Range
+
 
 class HttpBaseHandler:
     def __init__(self, match_criteria: Dict[str, List], context: Dict, server_callback: Callable = None):
@@ -132,31 +135,3 @@ class LoadBalancingHandler(ReverseProxyHandler):
         remote_host, remote_port = strategy_func()
         self.use_server_callback(**{f'{remote_host}:{remote_port}':1})
         return self.connect_and_send(remote_host, remote_port)    
-
-class ManageHandlers:
-    """
-    picks the handler based on settings and injects the needed context and the matching criteria for each handler. A
-    callback can additionally be passed in so that any handler that needs it (for updating fields in the server class, for example)
-    can get it. This callback is currently only used to update statistics
-    """
-
-    def __init__(self, settings: Dict, callback_to_attach: Callable = None):
-        self.tasks = settings['tasks']
-        self.callback = callback_to_attach
-        self.implemented_handlers = {
-            'serve_static':StaticAssetHandler,
-            'reverse_proxy':ReverseProxyHandler,
-            'load_balance':LoadBalancingHandler,
-            'health_check':HealthCheckHandler}
-    
-    def prepare_handlers(self) -> List[HttpBaseHandler]:
-        task_handlers: List[HttpBaseHandler] = []
-        for task_name, task_info in self.tasks.items():
-            match_criteria = task_info['match_criteria']
-            needed_context = task_info['context']
-            if task_name in self.implemented_handlers:
-                handler_class = self.implemented_handlers[task_name]
-                task_handlers.append(handler_class(match_criteria, needed_context, self.callback))
-            else:
-                raise NotImplementedError
-        return task_handlers
